% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions and examples.R
\name{AllelicMR_est}
\alias{AllelicMR_est}
\title{Function of Allele-Specific Mendelian Expression Estimation}
\usage{
AllelicMR_est(exprs.list)
}
\arguments{
\item{exprs.list, }{List of expression: X1, X2 are allele specific expression level,
X is total expression level, Y is the outcome, Z1 and Z2 are genotypes.}
}
\value{
Estimated parameters, including causal effect.
}
\description{
Estimate allele-specific Mendelian randomization parameters from expression
}
\examples{
Sim.out = sim_allelicMR()
Est.AMR = AllelicMR_est(Sim.out$Y, Sim.out$X, Sim.out$X1, Sim.out$X2, Sim.out$Z1, Sim.out$Z2)
Est.AMR
}
\seealso{
\code{\link{feasible.hin}},  \code{\link{VarCov}},
\code{\link{Likelihood.beta}},
}
