% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions and examples.R
\name{feasible.hin}
\alias{feasible.hin}
\title{Function of feasible starting point}
\usage{
feasible.hin(theta)
}
\arguments{
\item{theta, }{possible parameters for likelihood estimation}
}
\value{
adjusted feasible parameters.
}
\description{
Transform possible parameters to feasible area
}
