% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions and examples.R
\name{Likelihood.1}
\alias{Likelihood.1}
\title{Function for calculating log-likelihood, stage 1}
\usage{
Likelihood.1(theta, exprs.list)
}
\arguments{
\item{theta, }{first stage parameters, including sigma.e, sigma.u, sigma.w, mu.u and mu.w.}

\item{exprs.list, }{expression list, but Y is not used in this function;}
}
\value{
Log-likelihoodof the first stage.
}
\description{
This function calculate the log-likelihood of first stage
\code{X = 2*U + Z1*W1 + Z2*W2 + e} for homozygous individuals and
allele specific expression \code{X1 = U + Z1*W1 + e1;}
\code{X2 = Z2*W2 + e2} for heterozygous individuals.
}
