% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions and examples.R
\name{Likelihood.2}
\alias{Likelihood.2}
\title{Function for calculating log-likelihood of second stage
This function take estimated parameters from first stage, then estimate the
second stage parameters sigma.v, rho, mu.v and beta by maxizing full log-likelihood.}
\usage{
Likelihood.2(theta, theta.x, exprs.list)
}
\arguments{
\item{theta, }{parameters for second stage;}

\item{theta.x, }{estimated parameters from first stage;}

\item{exprs.list, }{List of expressions: total expression X, allele specific expression X1 and X2,
genotype Z1, Z2, and outcome Y;}
}
\value{
Log-likelihood
}
\description{
Function for calculating log-likelihood of second stage
This function take estimated parameters from first stage, then estimate the
second stage parameters sigma.v, rho, mu.v and beta by maxizing full log-likelihood.
}
