% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions and examples.R
\name{Likelihood.beta}
\alias{Likelihood.beta}
\title{Function for calculating log-likelihood}
\usage{
Likelihood.beta(theta, exprs.list)
}
\arguments{
\item{theta, }{all parameters;}

\item{exprs.list, }{List of outcome Y, allele specific expression X1, X2, and total expression X, and genotype Z1, Z3;}
}
\value{

}
\description{
Function for calculating log-likelihood
}
