% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions and examples.R
\name{sim_allelicMR}
\alias{sim_allelicMR}
\title{Function of simulation allele specific expression from normal distribution}
\usage{
sim_allelicMR(n = 1000, maf = 0.1, mean.u = 1, sd.u = 1, mean.v = 1,
  sd.v = 1, mean.w = 1, sd.w = 1, sd.e = sqrt(2)/4, rho = 0.5,
  beta = 1)
}
\arguments{
\item{n, }{number of samples, default value is 1000;}

\item{maf, }{minor allele frequency, default at 0.1;}

\item{mean.u, }{mean of U, default at 1;}

\item{sd.u, }{standard deviation of U, default at 1;}

\item{mean.v, }{mean of V, default at 1;}

\item{sd.v, }{standard deviation of V, default at 1;}

\item{mean.w, }{mean of W, default at 1;}

\item{sd.w, }{standard deviation of W, default at 1;}

\item{sd.e, }{standard deviation of measurement error, default at sqrt(2)/4;}

\item{rho, }{correlation between U and V, default at 0.5;}

\item{beta, }{causal effect size, default at 1;}
}
\value{
a list of allele specific expression: Y, X, X1, X2, Z1, Z2
}
\description{
Function of simulation allele specific expression from normal distribution
}
\details{
Data are generated with model
\code{Z1} and \code{Z2} are genotypes, take values on 0 or 1.
\code{Z1 = rbinom(n, 1, maf)}
\code{Z1 + Z2 = 1} are heterozygous indiviudals, we can observe allele specific
expression \code{X1} and \code{X2}; \code{Z1 + Z2 = 0} or \code{2} are homozygous
individuals, we can only observe total expression \code{X}. For heterozygous
individuals,
\code{X1 = U + Z1*W1 + e1; T2 = U + Z2*W2 + e2; X = X1 + X2;}
and for homozygous individuals, we can observe
\code{X = 2U + Z1*W1 + Z2*W2 + e;}
where \code{W1} and \code{W2} are independent effect of genotypes.
\code{W1 = rnorm(n, mean.w, sd.w); W2 = rnorm(n, mean.w, sd.w);}
Outcome \code{Y = V + beta* (2U + Z1*W1 + Z2*W2)}, where \code{beta} is the causal
effect and \code{V} are confounder effect,
\code{(U, V) = mvrnorm(n, mu = c(mean.u, mean.v), Sigma = matrix(c(sd.u^2, rho*sd.u*sd.v, rho*sd.u*sd.v, sd.v^2),nrow = 2))}
}
\examples{
library(MASS)
Sim.out = sim_allelicMR()
str(Sim.out)
}
