% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions and examples.R
\name{tsMLE_est}
\alias{tsMLE_est}
\title{Function of two-stage maximize likelihood estimation}
\usage{
tsMLE_est(exprs.list)
}
\arguments{
\item{exprs.list}{List of outcome Y, allele specific expression X1, X2, and total expression X, and genotype Z1, Z3;}
}
\value{
causal effect and other parameters
}
\description{
Two-stage MLE
}
