% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions and examples.R
\name{VarCov.1}
\alias{VarCov.1}
\title{Function for calculating possible starting point for first stage parameters}
\usage{
VarCov.1(exprs.list)
}
\arguments{
\item{exprs.list, }{List of outcome Y, allele specific expression X1, X2, and total expression X, and genotype Z1, Z3;}

\item{Beta, }{possible estimated beta;}
}
\value{
starting points theta;
}
\description{
Calculate possible starting point for first stage parameters
}
\examples{
Sim.out = sim_allelic()
theta = VarCov.1(Sim.out)
}
