% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions and examples.R
\name{VarCov}
\alias{VarCov}
\title{Function for calculating possible starting point for all parameters}
\usage{
VarCov(Beta, exprs.list)
}
\arguments{
\item{Beta, }{possible estimated beta;}

\item{exprs.list, }{List of outcome Y, allele specific expression X1, X2, and total expression X, and genotype Z1, Z3;}
}
\value{
starting points theta;
}
\description{
Calculate possible starting point for all parameters
}
\examples{
Sim.out = sim_allelic()
Xz = lm(X~Z)$fitted
Yz = lm(Y~Z)$fitted
beta.tsls = lm(Yz ~ Xz)$coef[2] # Two-stage least squares
theta = VarCov(beta.tsls, Sim.out)
}
